# openapi/climate-hub-v1.yaml
openapi: "3.0.3"
info:
  title: Climate Hub API — Need Engine Closed Loop
  version: "1.1.0"
  description: |
    REST API локальной платформы управления микроклиматом здания.
    Реализован замкнутый контур Need Engine: Environment → Need → Capability →
    Device → Command → Effect → Satisfaction.
servers:
  - url: /api/v1
    description: Climate Hub API
paths:
...
  /buildings/{buildingId}:
    get:
      operationId: getBuilding
      summary: Информация о здании
      tags: [Buildings]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Данные здания
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /buildings/{buildingId}/floors:
    get:
      operationId: getBuildingFloors
      summary: Этажи здания с помещениями
      tags: [Buildings]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Список этажей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FloorSummary"

  # ────────────────────── Floors / Rooms ──────────────────────
  /floors/{floorId}/rooms:
    get:
      operationId: getFloorRooms
      summary: Список помещений этажа
      tags: [Rooms]
      parameters:
        - name: floorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Список помещений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomSummary"
  /rooms/{roomId}:
    get:
      operationId: getRoom
      summary: Информация о помещении
      tags: [Rooms]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Данные помещения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSummary"
        "404":
          $ref: "#/components/responses/NotFound"

  # ────────────────────── Policy ──────────────────────
...
  /devices:
    get:
      operationId: listDevices
      summary: Реестр устройств
      tags: [Devices]
      responses:
        "200":
          description: Список устройств
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceSummary"
    post:
      operationId: registerDevice
      summary: Регистрация устройства
      tags: [Devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: Устройство зарегистрировано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSummary"
  /devices/{deviceId}:
    get:
      operationId: getDevice
      summary: Карточка устройства
      tags: [Devices]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Данные устройства
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /devices/{deviceId}/assignments:
    post:
      operationId: assignDevice
      summary: Назначить устройство в помещение
      tags: [Devices]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roomId:
                  type: string
                  format: uuid
              required: [roomId]
      responses:
        "200":
          description: Устройство назначено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSummary"
  /rooms/{roomId}/devices:
    get:
      operationId: getRoomDevices
      summary: Устройства помещения
      tags: [Devices]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Список устройств помещения
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceSummary"

  # ────────────────────── Commands ──────────────────────
...
  /needs/buildings/{buildingId}:
    get:
      operationId: getNeedsByBuilding
      summary: Потребности здания
      tags: [Needs]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Потребности здания
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeedResponse"
...
components:
  schemas:
...
    ControlMode:
      type: string
      enum: [disabled, monitorOnly, manual, automatic]

    # ── Needs ──
    NeedType:
      type: string
      enum: [temperatureHeating, temperatureCooling, humidityIncrease, humidityDecrease, co2Reduction, illuminanceIncrease, illuminanceDecrease]
    NeedStatus:
      type: string
      enum: [detected, planning, planned, executing, waitingForEffect, satisfied, blocked, cancelled, expired]
    NeedSeverity:
      type: string
      enum: [low, medium, high, critical]
    NeedResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/NeedType"
        status:
          $ref: "#/components/schemas/NeedStatus"
        severity:
          $ref: "#/components/schemas/NeedSeverity"
        controlMode:
          $ref: "#/components/schemas/ControlMode"
        currentValue:
          type: number
          nullable: true
        deviation:
          type: number
        desiredMin:
          type: number
        desiredMax:
          type: number
        desiredPreferred:
          type: number
        sourceParameterCode:
          type: string
          nullable: true
        selectedCapabilityCode:
          type: string
          nullable: true
        selectedDeviceId:
          type: string
          format: uuid
          nullable: true
        activeCommandId:
          type: string
          format: uuid
          nullable: true
        planningFailureCode:
          type: string
          nullable: true
        violationSince:
          type: string
          format: date-time
          nullable: true
        stableSince:
          type: string
          format: date-time
          nullable: true
        cooldownUntil:
          type: string
          format: date-time
          nullable: true
        effectEvaluationDueAt:
          type: string
          format: date-time
          nullable: true
        lastEvaluationAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
...
    BuildingSummary:
      type: object
      required:
        - id
        - name
        - status
        - floorsCount
        - roomsCount
        - roomsWithWarnings
        - devicesOnline
        - devicesOffline
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [normal, warning, critical, offline, unknown]
        floorsCount:
          type: integer
        roomsCount:
          type: integer
        roomsWithWarnings:
          type: integer
        devicesOnline:
          type: integer
        devicesOffline:
          type: integer
        updatedAt:
          type: string
          format: date-time
    FloorSummary:
      type: object
      required:
        - id
        - buildingId
        - name
        - number
        - rooms
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        name:
          type: string
        number:
          type: integer
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/RoomSummary"
    RoomSummary:
      type: object
      required:
        - id
        - floorId
        - name
        - status
        - environment
        - activeWarningsCount
        - devicesOnline
        - devicesTotal
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        floorId:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [normal, warning, critical, offline, stale, unknown]
        environment:
          $ref: "#/components/schemas/RoomEnvironmentSummary"
        activeWarningsCount:
          type: integer
        devicesOnline:
          type: integer
        devicesTotal:
          type: integer
        updatedAt:
          type: string
          format: date-time
    RoomEnvironmentSummary:
      type: object
      properties:
        temperatureC:
          type: number
          nullable: true
        relativeHumidityPct:
          type: number
          nullable: true
        co2Ppm:
          type: number
          nullable: true
        quality:
          $ref: "#/components/schemas/MeasurementQuality"
    MeasurementQuality:
      type: string
      enum: [valid, estimated, stale, outOfRange, sensorError, rejected, unknown]
...
    DeviceSummary:
      type: object
      required:
        - id
        - hardwareId
        - name
        - model
        - deviceType
        - status
        - connectivity
        - capabilities
      properties:
        id:
          type: string
          format: uuid
        hardwareId:
          type: string
        name:
          type: string
        model:
          type: string
        deviceType:
          type: string
        status:
          type: string
          enum: [registered, active, inactive, decommissioned]
        connectivity:
          type: string
          enum: [online, offline, suspected, synchronizing, operational]
        firmwareVersion:
          type: string
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        assignedRoomId:
          type: string
          format: uuid
          nullable: true
        assignedRoomName:
          type: string
          nullable: true
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/DeviceCapability"
    DeviceCapability:
      type: object
      required:
        - code
        - kind
        - writable
        - available
      properties:
        code:
          type: string
        kind:
          type: string
          enum: [measurement, control, system]
        unit:
          type: string
        minimum:
          type: number
          nullable: true
        maximum:
          type: number
          nullable: true
        writable:
          type: boolean
        available:
          type: boolean
    RegisterDeviceRequest:
      type: object
      required:
        - hardwareId
        - name
        - manufacturer
        - modelName
        - protocolVersion
      properties:
        hardwareId:
          type: string
        name:
          type: string
        manufacturer:
          type: string
        modelName:
          type: string
        hardwareVersion:
          type: string
        protocolVersion:
          type: string
    ApiProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errorCode:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiProblemDetails"
...